Refactoring JavaClient experiment - with corrected build.xml
